Fixed an incompatible type warning

+++ include/msgpack/fbuffer.h
@@ -23,7 +23,7 @@ extern "C" {
  * @{
  */
 
-static inline int msgpack_fbuffer_write(void* data, const char* buf, unsigned int len)
+static inline int msgpack_fbuffer_write(void* data, const char* buf, size_t len)
 {
     return (1 == fwrite(buf, len, 1, (FILE *)data)) ? 0 : -1;
 }
